+++ Fossil/src/Game.h
@@ -7,11 +7,17 @@
 class FSL_API Game
 {
 public:
+	bool running;
+
 	Game();
 	virtual ~Game();
 
 	void InitializeGame();
 	void Run();
+	void Update();
+	void Display();
+	
+	int GetTickCount();
 };
 
 //	Definined in client