+++ Workspace/src/Project.cpp
@@ -1,7 +1,5 @@
 #include "Fossil.h"
 
-Log logger2;
-
 class Workspace : public Game
 {
 public:
@@ -11,21 +9,10 @@ public:
 
 	~Workspace()
 	{
-
 	}
 };
 
-int main()
-{
-
-	//	Log message noting that the game started
-	logger2.logMessage(". . . Starting . . . \n", "gamelogs.txt", true);
-
-	//	Create instance of game
-	Workspace game = Workspace();
-	game.InitializeGame();
-	game.Run();
-
-	//	Log message noting that the game has exited
-	logger2.logMessage("\n . . . Ending . . .", "gamelogs.txt", false);
+//	Create new game pointer
+Game* CreateGame() {
+	return new Workspace();
 }